Store custom errors in a separate file
